# src/main/resources/swagger/swagger.json
{
  "swagger":"2.0",
  "info":{
    "version":"1.0",
    "title":"Sample Product API"
  },
  "basePath":"/",
  "tags":[
    {
      "name":"product"
    }
  ],
  "paths":{
    "/product/{product}":{
      "get":{
        "tags":[
          "product"
        ],
        "summary":"Find a product by ID",
        "description":"Returns a product with the provided ID if exists",
        "operationId":"getProduct",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"product",
            "in":"path",
            "description":"ID of product that needs to be fetched",
            "required":true,
            "type":"string",
            "maximum":1000.0,
            "minimum":1.0
          }
        ],
        "responses":{
          "200":{
            "description":"successful operation",
            "schema":{
              "$ref":"#/definitions/Product"
            }
          },
          "400":{
            "description":"Invalid ID supplied"
          },
          "404":{
            "description":"Product not found"
          }
        }
      },
      "post":{
        "tags":[
          "product"
        ],
        "summary": "Create a Product",
        "description": "Creates a new product with product id",
        "operationId": "createProduct",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"product",
            "in":"path",
            "required":true,
            "type":"string"
          },
          {
            "in":"body",
            "name":"body",
            "required":false,
            "schema":{
              "$ref":"#/definitions/Product"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"successful operation",
            "schema":{
              "$ref":"#/definitions/Product"
            },
            "headers":{

            }
          }
        }
      },
      "put":{
        "tags":[
          "product"
        ],
        "operationId":"updateProduct",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"product",
            "in":"path",
            "required":true,
            "type":"string"
          },
          {
            "in":"body",
            "name":"body",
            "required":false,
            "schema":{
              "$ref":"#/definitions/Product"
            }
          }
        ],
        "responses":{
          "default":{
            "description":"successful operation"
          }
        }
      },
      "delete":{
        "tags":[
          "product"
        ],
        "summary": "Delete a product by ID",
        "description": "Deletes a product with the provided ID if exists",
        "operationId":"deleteProduct",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"body",
            "name":"body",
            "required":false,
            "schema":{
              "$ref":"#/definitions/Product"
            }
          }
        ],
        "responses":{
          "200":{
            "description":"successful operation",
            "schema":{
              "type":"string"
            },
            "headers":{

            }
          }
        }
      }
    }
  },
  "definitions":{
    "Product":{
      "type":"object",
      "properties":{
        "productId":{
          "type":"string"
        },
        "productName":{
          "type":"string",
          "description":"Name of the product"
        },
        "category":{
          "type":"string"
        },
        "price":{
          "type":"string"
        }
      }
    }
  }
}